% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alcoholdependence}
\alias{alcoholdependence}
\title{Home safety data}
\format{A data frame with 87 rows and 12 variables. The variables are as
follows:

\itemize{
  \item studyid: study identification number
  \item study: study progressive number in the sample
  \item arm: single arm of a specific study
  \item trt: intervention in the specific study arm; 1 means the
  control/baseline treatment
  \item baseline: baseline treatment for a specific study
  \item narms: total number of arms for a specific study
  \item n1,n2,n3: sample sizes for each study arm
  \item y1,y2,y3: outcomes (number of events) for each study arm
}
  There are 4 interventions corresponding to:
\itemize{
  \item treatment 1: placebo
  \item treatment 2: naltrexone
  \item treatment 3: acamprosate
  \item treatment 4: naltrexone + acamprosate
}}
\source{
Liu, Y., DeSantis, S. M., and Chen, Y. (2018) Bayesian mixed treatment comparisons meta-analysis for correlated outcomes subject to reporting bias. Journal of the Royal Statistical Society C (67) 127–144
}
\usage{
alcoholdependence
}
\description{
A systematic review of pharmacological treatments for alcohol dependence, published in two consecutive Cochrane reviews.
}
\keyword{datasets}

