% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d_logprior}
\alias{d_logprior}
\title{Log-likelihood of copula based model for a multivariate NMA.}
\usage{
d_logprior(d, d_sigma, ref_trt)
}
\arguments{
\item{nc_data}{pippo}

\item{x}{pippo}

\item{mu}{pippo}

\item{delta}{pippo}

\item{Gamma}{pippo}
}
\value{
A length-one numeric vector.
}
\description{
Evaluation of the log-likelihood.
}
\examples{
# nothing for now!
}

