% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{delta_logpost}
\alias{delta_logpost}
\title{Log-likelihood of copula based model for a multivariate NMA.}
\usage{
delta_logpost(delta, mu, tau, eta, y, n, w, gamma, eps, eps_ab)
}
\arguments{
\item{delta}{pippo}

\item{mu}{pippo}

\item{nc_data}{pippo}

\item{x}{pippo}

\item{Gamma}{pippo}
}
\value{
A length-one numeric vector.
}
\description{
Evaluation of the log-likelihood.
}
\examples{
# nothing for now!
}

