% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{delta_logprior}
\alias{delta_logprior}
\title{Log-likelihood of copula based model for a multivariate NMA.}
\usage{
delta_logprior(delta, d, Sigma_M, trt_arms, baseline, narms)
}
\arguments{
\item{delta}{pippo}

\item{nc_data}{pippo}

\item{x}{pippo}

\item{mu}{pippo}

\item{Gamma}{pippo}
}
\value{
A length-one numeric vector.
}
\description{
Evaluation of the log-likelihood.
}
\examples{
# nothing for now!
}

