% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dinvgamma_rcpp}
\alias{dinvgamma_rcpp}
\title{Inverse gamma density.}
\usage{
dinvgamma_rcpp(x, alpha, beta, logd = FALSE)
}
\arguments{
\item{x}{Numeric vector whose representing the points at which to evaluate
the density.}

\item{alpha}{Inverse gamma shape parameter. Must be strictly positive.}

\item{beta}{Inverse gamma scale parameter. Must be strictly positive.}

\item{logd}{Boolean length-one vector; if TRUE the log density is returned.}
}
\value{
A numeric vector.
}
\description{
Evaluate the inverse gamma density.
}
\examples{
x <- seq(.01, 50, length.out = 1000)
alpha <- 1
beta <- 5
res <- dinvgamma_rcpp(x, alpha, beta)
plot(x, res, type = "l", main = "Inverse Gamma density")
}

