% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dinvwish_arma}
\alias{dinvwish_arma}
\title{Inverse Wishart density.}
\usage{
dinvwish_arma(IW, nu, S, logd = FALSE)
}
\arguments{
\item{IW}{Numeric matrix at which to evaluate the density.}

\item{nu}{Integer length-one vector providing the number of degrees of
freedom.}

\item{S}{Symmetric, positive definite numeric matrix representing the
distribution scale matrix.}

\item{logd}{Boolean length-one vector; if TRUE the log density is returned.}
}
\value{
A length-one numeric vector.
}
\description{
Evaluation of the inverse Wishart density.
}
\examples{
set.seed(123)
### Parameters
k <- 8
S <- bayesm::rwishart(10, diag(k))$IW
nu <- round(runif(1, min = (k + 1), max = 30))
### Benchmarking
n <- 1e+2
X <- array(NA, dim = c(n, k, k))
for (i in 1:n) X[i, , ] <- MCMCpack::riwish(nu, S)
require(rbenchmark)
benchmark(apply(X, 1, MCMCpack::diwish, nu, S),
          apply(X, 1, dinvwish_arma, nu, S, FALSE),
          order = "relative", replications = 100)[, 1:4]
}

