% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlkj_arma}
\alias{dlkj_arma}
\title{Compute the density function of a Lewandowski-Kurowicka-Joe distribution.}
\usage{
dlkj_arma(R, eta, logd = FALSE)
}
\arguments{
\item{R}{A numeric matrix that representing the correlation matrix.}

\item{eta}{Length-one numeric vector representing the parameter of the
distribution.}

\item{logd}{Boolean length-one vector; if TRUE the log density is returned.}
}
\value{
A length-one numeric vector.
}
\description{
Compute the density function of a Lewandowski-Kurowicka-Joe distribution
for a correlation matrix.
}
\details{
This function is a port of the corresponding Stan (\url{http://mc-stan.
org}) function developed using the Eigen library (\url{http://eigen.
tuxfamily.org}).
}
\examples{
set.seed(123)
K <- 5
R <- rlkj_arma(K, .5) # eta = 0.5
dlkj_arma(R, .5)
R <- rlkj_arma(K, 1) # eta = 1
dlkj_arma(R, 1)
R <- rlkj_arma(K, 2) # eta = 2
dlkj_arma(R, 2)
}
\references{
Lewandowski, D., Kurowicka, D., and Joe, H. (2009). Generating random
correlation matrices based on vines and extended onion method. Journal of
Multivariate Analysis, 100:1989-2001.
}
