% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlogchol_arma}
\alias{dlogchol_arma}
\title{Log Cholesky prior distribution.}
\usage{
dlogchol_arma(A, sigma_r, logd = FALSE)
}
\arguments{
\item{A}{A numeric matrix at which to evaluate the prior density.}

\item{sigma_r}{Length-one numeric vector providing the standard deviation
of the underlying normal distributions used to generate the Cholesky
factors.}
}
\value{
A length-one numeric vector.
}
\description{
Evaluation of the prior distribution based on the log Cholesky
factorization for a covariance matrix \eqn{\Sigma_M} .
}
\examples{
M <- 5
sigma_r <- 10^-(1/3)
rng <- round(runif(1, 1, 10000))
set.seed(rng)
Sigma <- rlogchol_arma(M, sigma_r)
dlogchol_arma(Sigma, sigma_r)
}
