% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmatvn_arma}
\alias{dmatvn_arma}
\title{Matrix variate normal density.}
\usage{
dmatvn_arma(X, M, U, V, logd = FALSE)
}
\arguments{
\item{X}{Numeric matrix at which to evaluate the density.}

\item{M}{Numeric matrix containing the univariate means.}

\item{U}{Numeric positive definite matrix representing the covariance
matrix among the X rows.}

\item{V}{Numeric positive definite matrix representing the covariance
matrix among the X columns.}

\item{logd}{Boolean length-one vector; if TRUE the log density is returned.}
}
\value{
A length-one numeric vector.
}
\description{
Evaluate the matrix variate normal density.
}
\examples{
set.seed(123)
n <- 10
p <- 3
x <- rnorm(n*p)
X <- matrix(x, nrow = n, ncol = p)
M <- matrix(0, nrow = n, ncol = p)
U <- diag(n)
V <- diag(p)
dmatvn_arma(X, M, U, V)
}
\references{
\url{https://en.wikipedia.org/wiki/Matrix_normal_distribution}
}

