% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmatvt_arma}
\alias{dmatvt_arma}
\title{Matrix variate t density.}
\usage{
dmatvt_arma(X, M, Sigma, Omega, df, logd = FALSE)
}
\arguments{
\item{X}{Numeric matrix at which to evaluate the density.}

\item{M}{Numeric matrix containing the univariate means.}

\item{Sigma}{Numeric positive definite matrix representing the covariance
matrix among the X columns.}

\item{Omega}{Numeric positive definite matrix representing the covariance
matrix among the X rows.}

\item{df}{Integer length-one vector providing the number of degrees of
freedom.}

\item{logd}{Boolean length-one vector; if TRUE the log density is returned.}
}
\value{
A length-one numeric vector.
}
\description{
Evaluate the matrix variate t density.
}
\examples{
set.seed(123)
n <- 10
p <- 3
x <- rnorm(n*p)
X <- matrix(x, nrow = n, ncol = p)
M <- matrix(0, nrow = n, ncol = p)
Sigma <- diag(n)
Omega <- diag(p)
df <- 5
dmatvt_arma(X, M, Sigma, Omega, df)
}
\references{
\url{https://en.wikipedia.org/wiki/Matrix_t_distribution}
}

