% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmvn_arma}
\alias{dmvn_arma}
\title{Multivariate normal density.}
\usage{
dmvn_arma(x, mean, sigma, logd = FALSE)
}
\arguments{
\item{x}{Numeric matrix whose rows represent the points at which to
evaluate the density.}

\item{mean}{Numeric vector containing the univariate means.}

\item{sigma}{Numeric positive definite matrix representing the covariance
matrix of the distribution.}

\item{logd}{Boolean length-one vector; if TRUE the log density is returned.}
}
\value{
A numeric vector.
}
\description{
Evaluate the multivariate normal density.
}
\examples{
set.seed(123)
### Covariance matrix and mean vector
sigma <- bayesm::rwishart(10, diag(8))$IW
mean <- rnorm(8)
### Benchmarking
n <- 1e+4
X <- mvtnorm::rmvnorm(n, mean, sigma)
require(rbenchmark)
benchmark(mvtnorm::dmvnorm(X, mean, sigma, log = FALSE), 
          dmvn_arma(X, mean, sigma, FALSE),
          order = "relative", replications = 100)[, 1:4]
}
\references{
\url{http://gallery.rcpp.org/articles/dmvnorm_arma/}
}

