% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmvt_arma}
\alias{dmvt_arma}
\title{Multivariate t density.}
\usage{
dmvt_arma(x, mean, sigma, df, logd = FALSE)
}
\arguments{
\item{x}{A numeric matrix whose rows represent the points at which to
evaluate the density.}

\item{mean}{A numeric vector containing the univariate means.}

\item{sigma}{A positive definite numeric matrix representing the covariance
matrix of the distribution.}

\item{df}{Integer length-one vector providing the number of degrees of
freedom.}

\item{logd}{Boolean length-one vector; if TRUE the log density is returned.}
}
\value{
A numeric vector.
}
\description{
Evaluate the multivariate t density.
}
\examples{
set.seed(123)
### Covariance matrix and mean vector
sigma <- bayesm::rwishart(10, diag(8))$IW
mean <- rnorm(8)
### Benchmarking
n <- 1e+4
df <- 20
X <- mvtnorm::rmvt(n, sigma, df, mean)
require(rbenchmark)
benchmark(mvtnorm::dmvt(X, mean, sigma, df, log = FALSE), 
          LearnBayes::dmt(X, mean, sigma, df, FALSE),
          dmvt_arma(X, mean, sigma, df, FALSE),
          order = "relative", replications = 100)[, 1:4]
}
\references{
\url{http://gallery.rcpp.org/articles/dmvnorm_arma/}
}
