% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dtruncnorm_rcpp}
\alias{dtruncnorm_rcpp}
\title{The truncated univariate normal distribution.}
\usage{
dtruncnorm_rcpp(x, a, b, mean, sd)
}
\arguments{
\item{x}{Numeric vector containing the values at which to evaluate the
density.}

\item{a}{Numeric vector providing the lower bounds. These may be
\code{-Inf}.}

\item{b}{Numeric vector providing the upper bounds. These may be
\code{Inf}.}

\item{mean}{Numeric vector of means.}

\item{sd}{Numeric vector of standard deviations.}
}
\value{
A numeric vector.
}
\description{
Density of a truncated normal distribution with mean equal to \code{mean}
and standard deviation equal to \code{sd}.
}
\details{
The values of \code{a}, \code{b}, \code{mean} and \code{sd} are recycled as
needed.

This function is a Rcpp port of the corresponding 
\code{\link[truncnorm]{dtruncnorm}} function from the \code{truncnorm}
package.
}
\examples{
a <- 0
b <- Inf
mean <- 0
sd <- 1
rng <- round(runif(1, 1, 10000))
set.seed(rng)
n <- 20
x <- rtruncnorm_rcpp(n, a, b, mean, sd)
dtruncnorm_rcpp(x, a, b, mean, sd)
}

