% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gamma_logpost}
\alias{Gamma_logpost}
\title{Log-likelihood of copula based model for a multivariate NMA.}
\usage{
Gamma_logpost(Gamma, x, eta)
}
\arguments{
\item{Gamma}{pippo}

\item{x}{pippo}

\item{nc_data}{pippo}

\item{mu}{pippo}

\item{delta}{pippo}
}
\value{
A length-one numeric vector.
}
\description{
Evaluation of the log-likelihood.
}
\examples{
# nothing for now!
}

