% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gausscopdens}
\alias{gausscopdens}
\title{Gaussian copula density evaluation.}
\usage{
gausscopdens(u, Gamma, is_u = FALSE, logd = FALSE)
}
\arguments{
\item{u}{A numeric vector providing the point at which to evaluate the
copula density.}

\item{Gamma}{A numeric matrix providing the Gaussian copula correlation
matrix. These elements need to be provided row-wise (i.e., g_12, g_13,
\ldots, g_1M, g_23, g_24,\ldots, g_2M,\ldots,g_(M-1)(M-1)).}

\item{is_u}{A length-one logical vector indicating whether the \code{u}
argument provides the \eqn{u} or the \eqn{\Phi^{-1}(u)} values.}

\item{logd}{Boolean length-one vector; if TRUE the log density is returned.}
}
\value{
A length-one numeric vector.
}
\description{
Evaluates a Gaussian copula density.
}
