% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc-package.R
\docType{data}
\name{homesafety}
\alias{homesafety}
\title{Home safety data}
\format{
A data frame with 45 rows and 12 variables. The variables are as
follows:

\itemize{
  \item studyid: study identification number
  \item study: study progressive number in the sample
  \item arm: single arm of a specific study
  \item trt: intervention in the specific study arm; 1 means the
  control/baseline treatment
  \item baseline: baseline treatment for a specific study
  \item narms: total number of arms for a specific study
  \item n1, n2, n3: sample sizes for each study arm
  \item y1, y2, y3: outcomes (number of events) for each study arm
}
  There are 9 interventions corresponding to:
\itemize{
  \item treatment 1: usual care
  \item treatment 2: education
  \item treatment 3: education + provision of free/low cost equipment
  \item treatment 4: education + provision of free/low cost equipment +
  home safety inspection
  \item treatment 5: education + provision of free/low cost equipment +
  fitting of equipment
  \item treatment 6: education + home safety inspection
  \item treatment 7: education + provision of free/low cost equipment +
  home safety inspection + fitting of equipment
  \item treatment 8: education + home visit
  \item treatment 9: provision of free/low cost equipment
}
}
\source{
\url{http://www.biomedcentral.com/1471-2288/14/92}
}
\usage{
homesafety
}
\description{
Data from a Cochrane systematic review of home safety education and
provision of safety equipment for injury prevention in children.
}
\keyword{datasets}
