% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{n_imputed}
\alias{n_imputed}
\title{Log-likelihood of copula based model for a multivariate NMA.}
\usage{
n_imputed(n_data)
}
\arguments{
\item{data}{pippo}

\item{init}{pippo}

\item{totiter}{pippo}

\item{prior}{pippo}

\item{prop}{pippo}

\item{verbose}{pippo}
}
\value{
A list containing the output of the MCMC simulation.
}
\description{
Evaluation of the log-likelihood.
}
\examples{
# nothing for now!
}

