% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_classes.R
\docType{class}
\name{nc_data-class}
\alias{nc_data-class}
\alias{nc_data}
\alias{initialize,nc_data-method}
\alias{show,nc_data-method}
\alias{summary,nc_data-method}
\title{An S4 class to represent a set of data for use in the \code{netcopula}
package.}
\usage{
\S4method{initialize}{nc_data}(.Object, ..., type, n_study, study_id,
  study_data, ref_trt)

\S4method{show}{nc_data}(object)

\S4method{summary}{nc_data}(object)
}
\arguments{
\item{.Object}{Prototype object from the class \code{\link{nc_data}}.}

\item{type}{A length-one character vector representing the data type (see
the corresponding slot description for more details).
studies (see the corresponding slot description for more details).}

\item{n_study}{A length-one numeric vector representing the number of
studies (see the corresponding slot description for more details).}

\item{study_id}{A three-columns data frame providing the studies ID (see the
corresponding slot description for more details).}

\item{study_data}{A data frame containing the outcome data for each study
(see the corresponding slot description for more details).}

\item{ref_trt}{A length-one integer vector providing the reference treatment
(which should not be confused with the baseline treatment for each study).}

\item{object}{An object of class \code{\link{nc_data}}.}

\item{\ldots}{Optional arguments to the function in its next call.}
}
\description{
An S4 class to represent a set of data for use in the \code{netcopula}
package.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initialize}: Create an instance of the \code{\link{nc_data}} class
using new/initialize.

\item \code{show}: Show an instance of the \code{\link{nc_data}} class.

\item \code{summary}: Provide a summary of a \code{\link{nc_data}} class
instance.
}}

\section{Slots}{

\describe{
\item{\code{type}}{A length-one character vector representing the data type and must
be either \code{"binary"}, \code{"count"} or \code{"continuous"}. Currently
the package supports only the "binary" data type.}

\item{\code{n_study}}{A length-one numeric vector representing the number of
studies.}

\item{\code{n_outcomes}}{A length-one numeric vector representing the number of
studies.}

\item{\code{n_datapoints}}{A length-one numeric vector representing the number of
data points in the sample (i.e., the overall number of arms for the trials
in the sample).}

\item{\code{n_treatments}}{A length-one numeric vector representing the total
number of treatments involved in the analysis.}

\item{\code{study_id}}{A four-columns data frame providing studies-specific data.
The four columns must be named \code{studyid}, \code{study}, \code{narms}
and \code{baseline}, with the second one that needs to be a progressive
number that match with those in the \code{study_data} slot. If NULL, a
default set of IDs will be created.}

\item{\code{study_data}}{A data frame containing the outcome data for each study.
Its first five columns must be named \code{study}, \code{arm}, \code{trt},
\code{baseline} and \code{narms}. Values in the \code{study} column must
match with those in the same column of the \code{study_id} slot. Outcomes
must occupy the last \code{n_outcomes} columns. Finally, the columns right
before the outcomes must provide the sample sizes for each study.}

\item{\code{ref_trt}}{A length-one integer vector providing the reference treatment
(which should not be confused with the baseline treatment for each study).}
}}

\examples{
showClass("nc_data")
getSlots("nc_data")
}
\author{
Sergio Venturini \email{sergio.venturini@unibocconi.it}

Sergio Venturini \email{sergio.venturini@unibocconi.it}

Sergio Venturini \email{sergio.venturini@unibocconi.it}

Sergio Venturini \email{sergio.venturini@unibocconi.it}
}
