% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_data_create.R
\name{nc_data_create}
\alias{nc_data_create}
\title{Create an instance of the \code{nc_data} class.}
\usage{
nc_data_create(type = character(1), n_study = numeric(1),
  n_outcomes = numeric(1), n_datapoints = numeric(1),
  n_treatments = numeric(1), study_id = data.frame(),
  study_data = data.frame(), ref_trt = integer(1))
}
\arguments{
\item{type}{A length-one character vector representing the data type and
must be either \code{"binary"}, \code{"count"} or \code{"continuous"}.}

\item{n_study}{A length-one numeric vector representing the number of
studies.}

\item{n_outcomes}{A length-one numeric vector representing the number of
studies.}

\item{n_datapoints}{A length-one numeric vector representing the number of
data points in the sample (i.e., the overall number of arms for the trials
in the sample).}

\item{n_treatments}{A length-one numeric vector representing the total
number of treatments involved in the analysis.}

\item{study_id}{A four-columns data frame providing studies-specific data.
The four columns must be named \code{studyid}, \code{study}, \code{narms}
and \code{baseline}, with the second one that needs to be a progressive
number that match with those in the \code{study_data} slot. If NULL, a
default set of IDs will be created.}

\item{study_data}{A data frame containing the outcome data for each study.
Its first five columns must be named \code{study}, \code{arm}, \code{trt},
\code{baseline} and \code{narms}. Values in the \code{study} column must
match with those in the same column of the \code{study_id} slot.}

\item{ref_trt}{A length-one integer vector providing the reference
treatment (which should not be confused with the baseline treatment for each
study).}
}
\value{
Returns an instance of the \code{\link{nc_data-class}}.
}
\description{
Create an instance of the \code{nc_data} class.
}
\examples{
data(homesafety)

n_study <- max(homesafety$study)
n_outcomes <- 3
n_datapoints <- nrow(homesafety)
n_treatments <- max(homesafety$trt)
studyid <- data.frame(studyid = unique(homesafety$studyid),
                      study = 1:n_study)

nc_data_create("binary", n_study, n_outcomes, n_datapoints, n_treatments,
               studyid, homesafety[, 2:12])
}

