% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_data_sim.R
\name{nc_data_missing}
\alias{nc_data_missing}
\title{Missing data simulation for the 'netcopula' package.}
\usage{
nc_data_missing(nc_data_list, n_miss = 5)
}
\arguments{
\item{nc_data_list}{List as generated by the \code{\link{nc_data_simulate}}
function.}

\item{n_miss}{A length-one numeric vector representing the number of missing
values.}
}
\value{
Returns a list with an instance of the \code{\link{nc_data-class}}
and the objects containing the parameters used to generate the data.
}
\description{
Function for adding missing values to a simulated data set. The function
currently generate missing values only according to the MCAR mechanism.
}
\examples{
set.seed(101)

n_outcomes <- 3
n_trt <- 3
n_studies <- 30
n_bin <- rep(100, n_outcomes) # binomial distributions sizes
d_init <- array(data = c(0, -0.5, -1, 0, 0.8, 0.3), dim = c(n_trt - 1, n_outcomes))
sigma2_vec <- c(1, 1.6, 1.8) # standard deviations in Sigma_M
rho_vec <- rep(.5, n_outcomes)
gamma_vec <- c(.3, .1, .5) # components of copula association matrix
mu_sigma2_sim <- 1e-1
ref_trt <- 1

nc_data_sim <- nc_data_simulate(d_init, sigma2_vec, rho_vec, gamma_vec, n_studies,
  n_trt, n_bin, n_outcomes, mu_sigma2_sim, ref_trt)
nc_data_sim <- nc_data_missing(nc_data_sim, n_miss = 10)
summary(nc_data_sim$nc_data)
}
