% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_data_create.R
\name{nc_data_simulate}
\alias{nc_data_simulate}
\title{Estimation of copula based models for multivariate network meta-analysis.}
\usage{
nc_data_simulate(d, sigma_vec, rho_vec, gamma_vec, n_studies, n_trt, n_bin,
  n_outcomes, mu_sigma2, ref_trt)
}
\arguments{
\item{data}{Object of class \code{nc_data} containing the data to analyze.}

\item{prm_init}{Named list of parameters for the proposal distributions of
the \eqn{Z} latent positions and \eqn{\alpha} parameters; the elements'
names must be \code{"z"} and \code{"alpha"} and both need to be a single
number representing the standard deviations of the normal proposal
distributions for each \eqn{z_{ih}^{g}}{z_ih^g} and \eqn{\alpha_g}
respectively.}

\item{random_start}{Length-one logical vector; if \code{TRUE} (default), it
allows the algorithm to start from a random starting point.}

\item{init}{Named list providing user-defined starting values. If provided,
it overrides the value given for \code{random_start}.}
}
\value{
\code{netcopula} returns an object of class \code{\link{nc_mcmc}}.
}
\description{
Estimation of copula based models for multivariate network meta-analysis.
}
\examples{
\dontrun{
# netcopula(...)
}

}
\author{
Sergio Venturini \email{sergio.venturini@unibocconi.it}
}
\references{
Venturini, S. and Graziani, R. (2016), "A Bayesian Copula Based Model for
Multivariate Network Meta-Analysis". Technical report.
}
\seealso{
\code{\link{nc_data-class}},
\code{\link{nc_mcmc}}
}

