% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nc_loglik}
\alias{nc_loglik}
\title{Log-likelihood of copula based model for a multivariate NMA.}
\usage{
nc_loglik(y, n, x, trt, mu, delta, Gamma)
}
\arguments{
\item{y}{pippo}

\item{n}{pippo}

\item{x}{pippo}

\item{trt}{pippo}

\item{mu}{pippo}

\item{delta}{pippo}

\item{Gamma}{pippo}
}
\value{
A length-one numeric vector.
}
\description{
Evaluation of the log-likelihood.
}
\examples{
# nothing for now!
}

