% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allclasses.R
\docType{class}
\name{nc_mcmc-class}
\alias{nc_mcmc-class}
\alias{plot,nc_mcmc-method}
\alias{summary,nc_mcmc-method}
\title{An S4 class to represent a MCMC run for a model in the \code{netcopula}
package.}
\usage{
\S4method{summary}{nc_mcmc}(object, digits_summary = 1, latent_var = FALSE,
  add_corr = FALSE, print = TRUE, ...)

\S4method{plot}{nc_mcmc}(x, y, latent_var = FALSE, add_corr = TRUE, to_plot,
  type = "trace", reorder = FALSE, ...)
}
\description{
An S4 class to represent a MCMC run for a model in the \code{netcopula}
package.
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: Provide a summary of a \code{\link{nc_mcmc}} class
instance.

\item \code{plot}: Provide a graphical summary of a \code{\link{nc_mcmc}}
class instance.
}}
\section{Slots}{

\describe{
\item{\code{x}}{BLA BLA BLA...}

\item{\code{delta}}{BLA BLA BLA...}

\item{\code{mu}}{BLA BLA BLA...}

\item{\code{Gamma}}{BLA BLA BLA...}

\item{\code{d}}{BLA BLA BLA...}

\item{\code{Sigma}}{BLA BLA BLA...}

\item{\code{accept}}{BLA BLA BLA...}

\item{\code{dens}}{BLA BLA BLA...}

\item{\code{control}}{BLA BLA BLA...}

\item{\code{dim}}{BLA BLA BLA...}

\item{\code{call}}{BLA BLA BLA...}
}}
\examples{
getSlots("nc_mcmc")
}

