% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nc_mcmc_mh}
\alias{nc_mcmc_mh}
\alias{netcopula-internal}
\alias{netcopula_internal}
\title{Internal functions for MCMC simulation.}
\usage{
nc_mcmc_mh(data, init, totiter, prior, prop, tuning, adapt, verbose)
}
\arguments{
\item{data}{internal data structure}

\item{init}{internal data structure}

\item{totiter}{internal data structure}

\item{prior}{internal data structure}

\item{prop}{internal data structure}

\item{tuning}{internal data structure}

\item{adapt}{internal data structure}

\item{verbose}{internal data structure}
}
\description{
For internal use only.
}
