% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{nc_mcmc_to_list}
\alias{nc_mcmc_to_list}
\title{Conversion of an \code{nc_mcmc} object to an object of class \code{list}.}
\usage{
nc_mcmc_to_list(res, include.burnin = FALSE, verbose = TRUE)
}
\arguments{
\item{res}{An object of type \code{nc_mcmc}.}

\item{include.burnin}{A logical scalar. If \code{TRUE} the burnin
iterations (if available) are not removed.}

\item{verbose}{A logical scalar. If \code{TRUE} prints additional
warnings during the conversion.}
}
\value{
An object of type \code{mcmc}.
}
\description{
\code{nc_mcmc_to_mcmc} converts an object of class \code{nc_mcmc}
  to an object with class \code{list}.
}
\seealso{
\code{\link{netcopula}()} for for fitting a copula-based multivariate
    network meta-analysis model;
  \code{\link{nc_mcmc-class}}.
}
\author{
Sergio Venturini \email{sergio.venturini@unito.it}
}
