% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_prior.R
\name{nc_prior}
\alias{nc_prior}
\alias{prior_nc}
\alias{check_prior}
\title{Auxiliary Function for Setting Model Priors.}
\usage{
nc_prior(
  mu_sigma2 = 10^3,
  d_sigma2 = 10^3,
  beta_sigma = 10^(-1/8),
  nGamma = 1,
  eta_prior = 1
)

prior_nc(
  mu_sigma2 = 10^3,
  d_sigma2 = 10^3,
  beta_sigma = 10^(-1/8),
  nGamma = 1,
  eta_prior = 1
)

check_prior(prior)
}
\arguments{
\item{mu_sigma2}{A length-one numeric vector representing the study-specific
baseline effects prior variance.
contain two numeric vectors, namely \code{a} and \code{b}.}

\item{d_sigma2}{A length-one numeric vector representing the pooled (across
trials) of treatment effects prior variance.}

\item{beta_sigma}{A length-one numeric vector representing the prior variance
of the Cholesky factorization factors for the between-study variance.}

\item{nGamma}{A length-one numeric vector representing the number of Gaussian
copula correlation matrices.}

\item{eta_prior}{A length-one numeric vector representing the Gaussian
copula correlation matrices prior parameter.}

\item{prior}{A named list of prior hyperparameters.}
}
\value{
A list with the prior hyperparameters as components.
}
\description{
{
\code{nc_prior()} is an auxiliary function as user interface for
  \code{netcopula()} fitting. Typically only used when calling the
  \code{netcopula()} function. It is used to set prior hyperparameters.

\code{prior_nc()} is an alias for \code{nc_prior()}.

\code{check_prior()} is an auxiliary function that verifies the
  correctness of the prior hyperparameters provided before a model is
  fitted with \code{\link{netcopula}()}.
}
}
\seealso{
\code{\link{netcopula}()}
}
\author{
Sergio Venturini \email{sergio.venturini@unibocconi.it}
}
\keyword{NMA}
