% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc-package.R
\docType{package}
\name{netcopula-package}
\alias{netcopula-package}
\alias{netcopula-pkg}
\title{Copula-Based Model for Multivariate Network Meta-Analysis.}
\description{
The \pkg{netcopula} package implements a Bayesian algorithm for fitting
a copula-based model for multivariate meta-analysis.

For efficiency reasons, the core computations in the package are implemented
using the \code{C} programming language and the \pkg{RcppArmadillo} package.

Plotting functionalities are imported from the nice \pkg{bayesplot} package.
Currently, the package includes methods for binary data only. In future
releases routines will be added specifically for continuous (i.e. normal),
and count data.
}
\section{\pkg{netcopula} classes}{

The \pkg{netcopula} package defines the following new classes:
\itemize{
  \item{\code{\link{nc_data}}: }{defines the data to use in a netcopula model.}
  \item{\code{\link{nc_mcmc}}: }{defines the results of a netcopula analysis
    for a single MCMC chain.}
}
The package includes \code{print}, \code{summary} and \code{plot} methods
  for each one of these classes.
}

\section{Resources}{

\itemize{
 \item{\strong{Bug reports}:}{
 If you have noticed a bug that needs to be fixed, please let us know at the
 \pkg{netcopula} issue tracker on GitHub:

 \url{https://github.com/sergioventurini/netcopula/issues/}.
 }
 \item{\strong{General questions and help}:}{
 To ask a question about \pkg{netcopula} send and email to:

 \email{sergio.venturini@sdabocconi.it}.
}
}
}

\references{
Graziani, R., Venturini, S. (2019), "Network Meta-Analysis for Adverse Events:
  An Exact (Bayesian) Approach Using Gaussian Copulas", Technical report.
}
\seealso{
\code{\link[bayesplot]{theme_default}} for the default ggplot theme
 used by \pkg{bayesplot}.

\code{\link[bayesplot]{bayesplot-colors}} to set or view the color
 scheme used for plotting with \pkg{bayesplot}.

\code{\link[ggplot2]{ggsave}} in \pkg{ggplot2} for saving plots.
}
