% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcopula.R
\docType{package}
\name{netcopula}
\alias{netcopula}
\alias{netcopula-package}
\title{\code{netcopula}: a package for estimating copula based models for
multivariate network meta-analysis (NMA).}
\usage{
netcopula(data, burnin = 10000, nsim = 5000, nthin = 1, prm_prior,
  prm_prop = NULL, prm_init, random_start = TRUE, init = NULL, tuning,
  adaptation, verbose = FALSE, mu_delta_mh = TRUE)
}
\arguments{
\item{data}{Object of class \code{nc_data} containing the data to analyze.}

\item{burnin}{Length one integer vector providing the number of MCMC burnin
iterations.}

\item{nsim}{Length one integer vector providing the number of MCMC
iterations.}

\item{nthin}{Length one integer vector providing the thinning interval.}

\item{prm_prior}{Named list of parameters for the proposal distributions of
the \eqn{Z} latent positions and \eqn{\alpha} parameters; the elements'
names must be \code{"z"} and \code{"alpha"} and both need to be a single
number representing the standard deviations of the normal proposal
distributions for each \eqn{z_{ih}^{g}}{z_ih^g} and \eqn{\alpha_g}
respectively.}

\item{prm_prop}{Named list of hyperparameters for the prior distributions
of the \eqn{\sigma^2} and \eqn{\lambda} parameters; the elements' names
must be \code{"sigma2"} and \code{"lambda"} with the first one a two-sized
vector containing the \eqn{\sigma^2} prior hyperparameters and the second
element a vector providing the \eqn{\lambda} hyperparameters.}

\item{random_start}{Length-one logical vector; if \code{TRUE} (default), it
allows the algorithm to start from a random starting point.}

\item{init}{Named list providing user-defined starting values. If provided,
it overrides the value given for \code{random_start}.}

\item{tuning}{Named list of tuning parameters (see details).}

\item{adaptation}{Named list of adaptation parameters (see details).}

\item{verbose}{Length-one logical vector; if \code{TRUE}, messages
informing about the progress of the computations will be printed.}
}
\value{
\code{netcopula} returns an object of class \code{\link{nc_mcmc}}.
}
\description{
The \code{netcopula} package provides methods for fitting copula based
models for network meta-analysis within a fully Bayesian framework.
Currently, it includes methods for binary data only.

Estimation of copula based models for multivariate network meta-analysis.
}
\examples{
\dontrun{
# netcopula(...)
}

}
\author{
Sergio Venturini \email{sergio.venturini@unibocconi.it}
}
\references{
Venturini, S. and Graziani, R. (2016), "A Bayesian Copula Based Model for
Multivariate Network Meta-Analysis". Technical report.
}
\seealso{
\code{\link{nc_data-class}},
\code{\link{nc_mcmc}}
}

