% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{param_long}
\alias{param_long}
\title{Internal functions.}
\usage{
param_long(prm_wide, narms, rowindex)
}
\arguments{
\item{prm_wide}{internal data structure}

\item{narms}{internal data structure}

\item{rowindex}{internal data structure}
}
\description{
For internal use only.
}
