% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{param_wide}
\alias{param_wide}
\title{Internal functions.}
\usage{
param_wide(prm_long, narms, trt, baseline)
}
\arguments{
\item{prm_long}{internal data structure}

\item{narms}{internal data structure}

\item{trt}{internal data structure}

\item{baseline}{internal data structure}
}
\description{
For internal use only.
}
