% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rinvgamma_rcpp}
\alias{rinvgamma_rcpp}
\title{Inverse gamma variates generation.}
\usage{
rinvgamma_rcpp(n, alpha, beta)
}
\arguments{
\item{n}{Numeric length-one vector providing the number of draws to
generate.}

\item{alpha}{Inverse gamma shape parameter. Must be strictly positive.}

\item{beta}{Inverse gamma scale parameter. Must be strictly positive.}
}
\value{
A numeric vector.
}
\description{
Generate of inverse gamma variates.
}
\examples{
set.seed(123)
n <- 1e4
alpha <- 3
beta <- 5
x <- sort(rinvgamma_rcpp(n, alpha, beta))
hist(x, breaks = 30, xlab = "x", freq = FALSE,
     main = "Inverse gamma variates")
lines(x, dinvgamma_rcpp(x, alpha, beta))
}

