% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rinvwish_arma}
\alias{rinvwish_arma}
\title{Inverse Wishart variates generation.}
\usage{
rinvwish_arma(nu, S)
}
\arguments{
\item{nu}{Integer length-one vector providing the number of degrees of
freedom.}

\item{S}{Symmetric, positive definite numeric matrix representing the
distribution scale matrix.}
}
\value{
A length-one numeric vector.
}
\description{
Generate a single inverse Wishart variate.
}
\examples{
set.seed(123)
### Parameters
k <- 8
S <- bayesm::rwishart(10, diag(k))$IW
nu <- round(runif(1, min = (k + 1), max = 30))
### Benchmarking
n <- 1e+2
require(rbenchmark)
benchmark(apply(rWishart(n, nu, S), 3, solve),
          replicate(n, MCMCpack::riwish(nu, S), simplify = "array"),
          replicate(n, rinvwish_arma(nu, S), simplify = "array"),
          columns = c("test", "replications", "relative", "elapsed"),
          order = "relative")
}
\references{
\url{https://en.wikipedia.org/wiki/
Wishart_distribution#Bartlett_decomposition}
}

