% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rlkj_arma}
\alias{rlkj_arma}
\title{Generate random variate from a Lewandowski-Kurowicka-Joe distribution.}
\usage{
rlkj_arma(K, eta)
}
\arguments{
\item{K}{Length-one numeric vector that represents the size of the matrix.}

\item{eta}{Length-one numeric vector representing the parameter of the
distribution.}
}
\value{
A numeric (correlation) matrix.
}
\description{
Generation of a single draw from a Lewandowski-Kurowicka-Joe distribution
for a correlation matrix.
}
\details{
This function is a port of the corresponding Stan (\url{http://mc-stan.
org}) function developed using the Eigen library (\url{http://eigen.
tuxfamily.org}).
}
\examples{
set.seed(123)
K <- 5
rlkj_arma(K, .5) # eta = 0.5
rlkj_arma(K, 1) # eta = 1
rlkj_arma(K, 2) # eta = 2
}
\references{
Lewandowski, D., Kurowicka, D., and Joe, H. (2009). Generating random
correlation matrices based on vines and extended onion method. Journal of
Multivariate Analysis, 100:1989-2001.
}
