% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmatvn_arma}
\alias{rmatvn_arma}
\title{Matrix variate normal generation.}
\usage{
rmatvn_arma(M, U, V)
}
\arguments{
\item{M}{Numeric matrix containing the univariate means.}

\item{U}{Numeric positive definite matrix representing the covariance
matrix among the X rows.}

\item{V}{Numeric positive definite matrix representing the covariance
matrix among the X columns.}
}
\value{
A numeric matrix.
}
\description{
Generate of a single matrix normal variate.
}
\examples{
set.seed(123)
n <- 10
p <- 3
M <- matrix(0, nrow = n, ncol = p)
U <- diag(n)
V <- diag(p)
rmatvn_arma(M, U, V)
set.seed(123)
rmvn_arma(n, rep(0, p), diag(p))
}
\references{
\url{https://en.wikipedia.org/wiki/Matrix_normal_distribution}
}

