% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmatvt_arma}
\alias{rmatvt_arma}
\title{Matrix variate t generation.}
\usage{
rmatvt_arma(M, Sigma, Omega, df)
}
\arguments{
\item{M}{Numeric matrix containing the univariate means.}

\item{Sigma}{Numeric positive definite matrix representing the covariance
matrix among the X columns.}

\item{Omega}{Numeric positive definite matrix representing the covariance
matrix among the X rows.}

\item{df}{Integer length-one vector providing the number of degrees of
freedom.}
}
\value{
A numeric matrix.
}
\description{
Generate of a single matrix t variate.
}
\examples{
set.seed(123)
n <- 10
p <- 3
x <- rnorm(n*p)
M <- matrix(0, nrow = n, ncol = p)
Sigma <- diag(n)
Omega <- diag(p)
df <- 5
rmatvt_arma(M, Sigma, Omega, df)
}
\references{
Kollo, T., van Rosen, D. (2005). Advanced Multivariate Statistics
with Matrices. Springer.
\url{https://en.wikipedia.org/wiki/Matrix_t_distribution}
}

