% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvn_arma}
\alias{rmvn_arma}
\title{Multivariate normal variates generation.}
\usage{
rmvn_arma(n, mean, sigma)
}
\arguments{
\item{n}{A length-one numeric vector providing the number of draws to
generate.}

\item{mean}{A numeric vector containing the univariate means.}

\item{sigma}{A positive definite numeric matrix representing the covariance
matrix of the distribution.}
}
\value{
A numeric matrix.
}
\description{
Generation of multivariate normal variates.
}
\examples{
set.seed(123)
### Covariance matrix and mean vector
sigma <- matrix(c(1, 0.9, -0.3, 0.9, 1, -0.4, -0.3, -0.4, 1), ncol = 3)
mu <- c(10, 5, -3)
### Benchmarking
n <- 1e+4
require(rbenchmark)
benchmark(mvtnorm::rmvnorm(n, mu, sigma),
          MASS::mvrnorm(n, mu, sigma),
          rmvn_arma(n, mu, sigma),
          columns = c("test", "replications", "relative", "elapsed"),
          order = "relative")
}
\references{
\url{http://gallery.rcpp.org/articles/simulate-multivariate-normal/}
}
