% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvt_arma}
\alias{rmvt_arma}
\title{Multivariate t variates generation.}
\usage{
rmvt_arma(n, mean, sigma, df)
}
\arguments{
\item{n}{Numeric length-one vector providing the number of draws to
generate.}

\item{mean}{Numeric vector containing the univariate means.}

\item{sigma}{Numeric positive definite matrix representing the covariance
matrix of the distribution.}

\item{df}{Integer length-one vector providing the number of degrees of
freedom.}
}
\value{
A numeric matrix.
}
\description{
Generate of multivariate t variates.
}
\examples{
set.seed(123)
### Covariance matrix and mean vector
sigma <- matrix(c(1, 0.9, -0.3, 0.9, 1, -0.4, -0.3, -0.4, 1), ncol = 3)
mu <- c(10, 5, -3)
df <- 7
### Benchmarking
n <- 1e+4
require(rbenchmark)
benchmark(mvtnorm::rmvt(n, sigma, df, mu),
          LearnBayes::rmt(n, mu, sigma, df),
          rmvt_arma(n, mu, sigma, df),
          columns = c("test", "replications", "relative", "elapsed"),
          order = "relative")
}
\references{
\url{http://gallery.rcpp.org/articles/simulate-multivariate-normal/}
}

