% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtruncnorm2_rcpp}
\alias{rtruncnorm2_rcpp}
\title{The truncated univariate normal distribution.}
\usage{
rtruncnorm2_rcpp(n, a, b, mean, sd)
}
\arguments{
\item{n}{Length-one numeric vector representing the number of draws.}

\item{a}{Numeric vector providing the lower bounds. These may be
\code{-Inf}.}

\item{b}{Numeric vector providing the upper bounds. These may be
\code{Inf}.}

\item{mean}{Numeric vector of means.}

\item{sd}{Numeric vector of standard deviations.}
}
\value{
A numeric vector.
}
\description{
Generation of draws from a truncated normal distribution with mean equal
to \code{mean} and standard deviation equal to \code{sd}.
}
\details{
The values of \code{a}, \code{b}, \code{mean} and \code{sd} are recycled as
needed.

This function is a Rcpp port of the corresponding 
\code{\link[bayesm]{rtrun}} function from the \code{bayesm}
package.
}
\examples{
n <- 20
a <- 0
b <- Inf
mean <- 0
sd <- 1
rng <- round(runif(1, 1, 10000))
set.seed(rng)
rtruncnorm2_rcpp(n, a, b, mean, sd)
}
\references{
Robert, C. (1995). Simulation of Truncated Normal Variables. Statistics
and Computing, 5, 121–125.
}

