% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtruncnorm_rcpp}
\alias{rtruncnorm_rcpp}
\title{The truncated univariate normal distribution.}
\usage{
rtruncnorm_rcpp(n, a, b, mean, sd)
}
\arguments{
\item{n}{Length-one numeric vector representing the number of draws.}

\item{a}{A numeric vector providing the lower bounds. These may be
\code{-Inf}.}

\item{b}{A numeric vector providing the upper bounds. These may be
\code{Inf}.}

\item{mean}{A numeric vector of means.}

\item{sd}{A numeric vector of standard deviations.}
}
\value{
A numeric vector.
}
\description{
Generation of draws from a truncated normal distribution with mean equal
to \code{mean} and standard deviation equal to \code{sd}.
}
\details{
The values of \code{a}, \code{b}, \code{mean} and \code{sd} are recycled as
needed.

This function is a Rcpp port of the corresponding 
\code{\link[truncnorm]{rtruncnorm}} function from the \code{truncnorm}
package.
}
\examples{
n <- 20
a <- 0
b <- Inf
mean <- 0
sd <- 1
rng <- round(runif(1, 1, 10000))
set.seed(rng)
rtruncnorm_rcpp(n, a, b, mean, sd)
}
\references{
Geweke, J. (1991). Efficient simulation from the multivariate normal and
student-t distributions subject to linear constraints. In Computing Science
and Statistics: Proceedings of the 23rd Symposium on the Interface, ed. E.
Keramidas and S. Kaufman, pp. 571-8. Fairfax Station, VA: Interface
Foundation of North America.
}
