% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwmh_adapt}
\alias{rwmh_adapt}
\title{Log-likelihood of copula based model for a multivariate NMA.}
\usage{
rwmh_adapt(theta, mu, rho, cov, ar, alpha, beta, gamma, tar, k, iter_cols, what)
}
\arguments{
\item{mu}{pippo}

\item{y}{pippo}

\item{n}{pippo}

\item{x}{pippo}

\item{trt}{pippo}

\item{delta}{pippo}

\item{Gamma}{pippo}
}
\value{
A length-one numeric vector.
}
\description{
Evaluation of the log-likelihood.
}
\examples{
# nothing for now!
}

