% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_classes.R
\docType{methods}
\name{show,nc_mcmc-method}
\alias{show,nc_mcmc-method}
\title{Show an instance of the \code{nc_mcmc} class.}
\usage{
\S4method{show}{nc_mcmc}(object)
}
\arguments{
\item{object}{An object of class \code{\link{nc_mcmc}}.}
}
\description{
Show an instance of the \code{nc_mcmc} class.
}
\author{
Sergio Venturini \email{sergio.venturini@unibocconi.it}
}
