% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sigma_block}
\alias{Sigma_block}
\title{Internal functions.}
\usage{
Sigma_block(Sigma_M, n)
}
\arguments{
\item{Sigma_M}{internal data structure}

\item{n}{internal data structure}
}
\description{
For internal use only.
}
