% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sigma_M_logpost}
\alias{Sigma_M_logpost}
\title{Log-likelihood of copula based model for a multivariate NMA.}
\usage{
Sigma_M_logpost(d, delta, Sigma_M, trt, baseline, narms_study, sigma_r)
}
\arguments{
\item{delta}{pippo}

\item{nc_data}{pippo}

\item{x}{pippo}

\item{mu}{pippo}

\item{Gamma}{pippo}
}
\value{
A length-one numeric vector.
}
\description{
Evaluation of the log-likelihood.
}
\examples{
# nothing for now!
}

