% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spearman_mcmc}
\alias{spearman_mcmc}
\title{Internal functions.}
\usage{
spearman_mcmc(Gamma_chain, n, M)
}
\arguments{
\item{Gamma_chain}{internal data structure}

\item{n}{internal data structure}

\item{M}{internal data structure}
}
\description{
For internal use only.
}
