% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gausscopdens}
\alias{gausscopdens}
\title{Gaussian copula density evaluation.}
\usage{
gausscopdens(u, Gamma, is_u, logd = FALSE)
}
\arguments{
\item{u}{A numeric vector providing the point at which to evaluate the
copula density.}

\item{Gamma}{A numeric matrix providing the Gaussian copula correlation
matrix. These elements need to be provided row-wise (i.e., g_12, g_13,
\ldots, g_1M, g_23, g_24,\ldots, g_2M,\ldots,g_(M-1)(M-1)).}

\item{is_u}{A length-one logical vector indicating whether the \code{u}
argument provides the \eqn{u} or the \eqn{\Phi^{-1}(u)} values.}

\item{logd}{Boolean length-one vector; if TRUE the log density is returned.}
}
\value{
A length-one numeric vector.
}
\description{
Evaluates a Gaussian copula density.
}
\examples{
u <- rep(0.1, 3)
Gamma <- c(.3, .5, .2)
gausscopdens(u, Gamma)

Gamma <- .3
len <- 100
u1 <- u2 <- seq(.01, .99, length.out = len)
dens <- matrix(NA, nrow = len, ncol = len)
for (i in 1:len) {
  for (j in 1:len) {
    dens[i, j] <- gausscopdens(c(u1[i], u2[j]), Gamma)
  }
}
persp(u1, u2, dens, theta = 120, phi = 25)
}
