% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_control.R
\name{nc_control}
\alias{nc_control}
\alias{control_nc}
\alias{check_control}
\title{Auxiliary Function for Controlling Model Fitting in the \code{netcopula } Package}
\usage{
nc_control(nsim = 50000, burnin = 1e+05, thin = 1,
  Gamma.update = "PX-RPMH", eta.prop = 0.5, sigma.r.prop = 0.01,
  random.start = TRUE, store.burnin = TRUE, verbose = FALSE)

control_nc(nsim = 50000, burnin = 1e+05, thin = 1,
  Gamma.update = "PX-RPMH", eta.prop = 0.5, sigma.r.prop = 0.01,
  random.start = TRUE, store.burnin = TRUE, verbose = FALSE)

check_control(control)
}
\arguments{
\item{nsim}{A length-one numeric vector for the number of draws to be taken
from the posterior distribution.}

\item{burnin}{A length-one numeric vector for the number of initial MCMC
iterations (usually to be discarded).}

\item{thin}{A length-one numeric vector for the number of iterations between
consecutive draws.}

\item{random.start}{A length-one logical vector. If \code{TRUE} the starting
values are drawn randomly, otherwise.}

\item{store.burnin}{A logical scalar. If \code{TRUE}, the samples from the
burnin are also stored and returned.}

\item{verbose}{A logical scalar. If \code{TRUE}, causes information to be
printed out about the progress of the fitting.}

\item{control}{A list of control options.}

\item{z.prop}{A length-one numeric vector providing the standard deviation of the
proposal distribution for the jump in the individual latent space
position.}

\item{alpha.prop}{A length-one numeric vector providing the standard deviation
of the proposal distribution for the jump in the individual random effect value.}
}
\value{
A named list with the control options as components.
}
\description{
{
\code{nc_control()} is an auxiliary function as user interface for
  \code{netcopula()} fitting. Typically only used when calling the \code{netcopula()}
  function. It is used to set parameters that affect the sampling but do
  not affect the posterior distribution.

\code{control_nc()} is an alias for \code{nc_control()}.

\code{check_control()} is an auxiliary function that verifies the
  correctness of the controls provided before a netcopula is fitted with
  \code{\link{netcopula}()}.
}
}
\examples{
\dontrun{
data(simdiss, package = "netcopula")
# Shorter run than default.
sim.fit <- netcopula(simdiss,
  control = nc_control(burnin = 1000, nsim = 2000, thin = 5, verbose = TRUE))
}

}
\seealso{
\code{\link{netcopula}()}
}
\author{
Sergio Venturini \email{sergio.venturini@unibocconi.it}
}
\keyword{based}
\keyword{clustering}
\keyword{model}
