% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_data_sim.R
\name{nc_data_simulate}
\alias{nc_data_simulate}
\title{Data simulation for the 'netcopula' package.}
\usage{
nc_data_simulate(d, sigma2_vec, rho_vec, gamma_vec, n_studies, n_trt,
  n_bin, n_outcomes, mu_sigma2, ref_trt)
}
\arguments{
\item{d}{Numeric matrix with dimension (#treatement - 1, #outcomes) containing
the true treatment effects used to simulate the data}

\item{sigma2_vec}{Numeric vector containing the true common between-study
variances.}

\item{rho_vec}{Numeric vector containing the true common between-study
correlations.}

\item{gamma_vec}{Numeric vector containing the true common Gaussian copula
correlation parameters.}

\item{n_studies}{A length-one numeric vector representing the number of
studies.}

\item{n_trt}{A length-one numeric vector representing the total number
of treatments involved in the analysis.}

\item{n_bin}{A length-one numeric vector representing the size of the
binomial distribution to sample the event numbers from.}

\item{n_outcomes}{A length-one numeric vector representing the number of
outcomes.}

\item{mu_sigma2}{A length-one numeric vector representing the variance of
the mu parameters (study-specific baseline effects).}

\item{ref_trt}{A length-one integer vector providing the reference
treatment (which should not be confused with the baseline treatment for each
study).}
}
\value{
Returns a list with an instance of the \code{\link{nc_data-class}}
and the objects containing the parameters used to generate the data.
}
\description{
Function for simulating data to use with the 'netcopula' package. The function
simulates data for the case of two arms studies, with the first half comparing
treatment 1 (reference) and 2 and the second half treatment 2 and 3.
}
\examples{
set.seed(101)

n_outcomes <- 3
n_trt <- 3
n_studies <- 30
n_bin <- rep(100, n_outcomes) # binomial distributions sizes
d_init <- array(data = c(0, -0.5, -1, 0, 0.8, 0.3), dim = c(n_trt - 1, n_outcomes))
sigma2_vec <- c(1, 1.6, 1.8) # standard deviations in Sigma_M
rho_vec <- rep(.5, n_outcomes)
gamma_vec <- c(.3, .1, .5) # components of copula association matrix
mu_sigma2_sim <- 1e-1
ref_trt <- 1

nc_data_sim <- nc_data_simulate(d_init, sigma2_vec, rho_vec, gamma_vec, n_studies,
  n_trt, n_bin, n_outcomes, mu_sigma2_sim, ref_trt)
summary(nc_data_sim$nc_data)
}
