% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_init.R
\name{nc_init}
\alias{nc_init}
\title{Function to compute the starting values before fitting a models with the
  \code{netcopula} package.}
\usage{
nc_init(
  data,
  prm_init,
  random_start,
  init = NULL,
  prm_prior = NULL,
  prm_prop = NULL,
  tuning = NULL,
  adaptation = NULL
)
}
\arguments{
\item{data}{An object of class \code{link{nc_data-class}} containing the data
to analyze.}

\item{prm_init}{An object of class \code{list} containing the parameters to
use for initialization.}

\item{random_start}{A length-one logical vector. If \code{TRUE} the starting
values are drawn randomly, otherwise.}

\item{init}{Either a length-one character vector or a named \code{list}
\code{list} providing user-defined starting values.}

\item{prm_prior}{An object of class \code{list} containing the prior
hyperparameters.}

\item{prm_prop}{An object of class \code{list} containing the proposal
parameters.}

\item{tuning}{An object of class \code{list} containing tuning parameters
used in the MCMC simulation.}

\item{adaptation}{An object of class \code{list} containing the parameters
used for adaptation of the MCMC simulation.}
}
\value{
A named \code{list} with the following items:
  \describe{
    \item{\code{mu}: }{array of latent coordinates starting values}
    \item{\code{delta}: }{numeric vector of initial cluster memberships}
    \item{\code{d}: }{numeric vector of initial cluster sizes}
    \item{\code{Sigma_M}: }{numeric vector of alpha starting values}
    \item{\code{Gamma}: }{numeric vector of eta starting values}
    \item{\code{D}: }{numeric vector of sigma2 starting values}
    \item{\code{x}: }{numeric vector of lambda starting values}
  }
}
\description{
\code{nc_init()} is the function that computes the initialization values for
  fitting a model with the \code{netcopula} package.
}
\seealso{
\code{\link{netcopula}()} for fitting a model with the
  \code{netcopula} package.
}
\author{
Sergio Venturini \email{sergio.venturini@unibocconi.it}
}
