% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_classes.R
\docType{class}
\name{nc_mcmc-class}
\alias{nc_mcmc-class}
\alias{nc_mcmc}
\title{An S4 class to represent a MCMC run for a model in the \code{netcopula}
package.}
\description{
An S4 class to represent the results of fitting the model in the
  \code{netcopula} package using a single Markov Chain Monte Carlo chain.
}
\section{Slots}{

\describe{
\item{\code{x}}{An object of class \code{array}; posterior draws from
the MCMC algorithm for the (adjusted) latent variables.}

\item{\code{x_unadj}}{An object of class \code{array}; posterior draws from
the MCMC algorithm for the (unadjusted) latent variables.}

\item{\code{delta}}{An object of class \code{array}; posterior draws from
the MCMC algorithm for the trial-specific log-odds ratios relative to the
baseline.}

\item{\code{mu}}{An object of class \code{array}; posterior draws from
the MCMC algorithm for the study-specific baseline effects.}

\item{\code{Gamma}}{An object of class \code{array}; posterior draws from
the MCMC algorithm for the copula correlation matrix.}

\item{\code{r}}{An object of class \code{array}; posterior draws from
the MCMC algorithm for the copula correlation matrix elements (used only
if Gamma.update is set to 'JointR').}

\item{\code{d}}{An object of class \code{array}; posterior draws from
the MCMC algorithm for the pooled effects (across trials) of each
treatment relative to the baseline.}

\item{\code{Sigma}}{An object of class \code{array}; posterior draws from
the MCMC algorithm for the between-study covariance structure.}

\item{\code{accept}}{An object of class \code{list}; final acceptance rates
for the MCMC algorithm.}

\item{\code{dens}}{An object of class \code{list}; list of log-likelihood,
log-prior and log-posterior values at each iteration of the MCMC
simulation.}

\item{\code{control}}{An object of class \code{list}; list of the control
parameters (number of burnin and sample iterations, etc.). See
\code{\link{nc_control}()} for more information.}

\item{\code{prior}}{An object of class \code{list}; list of the prior
hyperparameters. See \code{\link{nc_prior}()} for more information.}

\item{\code{dim}}{An object of class \code{list}; list of dimensions for
the estimated model.}

\item{\code{data}}{An object of class \code{link{nc_data-class}}; data used in the
analysis.}

\item{\code{call}}{An object of class \code{call} providing the matched call.}
}}

\examples{
getSlots("nc_mcmc")
}
