% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nc_mcmc_mh}
\alias{nc_mcmc_mh}
\title{Log-likelihood of copula based model for a multivariate NMA.}
\usage{
nc_mcmc_mh(data, init, totiter, prior, prop, tuning, adapt, verbose)
}
\arguments{
\item{data}{pippo}

\item{init}{pippo}

\item{totiter}{pippo}

\item{prior}{pippo}

\item{prop}{pippo}

\item{verbose}{pippo}
}
\value{
A list containing the output of the MCMC simulation.
}
\description{
Evaluation of the log-likelihood.
}
\examples{
# nothing for now!
}

