% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{netcopula}
\alias{netcopula}
\title{Estimation of a Copula-Based Model for Multivariate Network Meta-Analysis.}
\usage{
netcopula(
  data,
  control = nc_control(),
  prior,
  prm_init,
  init = NULL,
  tuning,
  adaptation
)
}
\arguments{
\item{data}{Object of class \code{nc_data} containing the data to analyze.}

\item{control}{A list of control parameters that affect the sampling
but do not affect the posterior distribution. See
\code{\link{nc_control}()} for more details.}

\item{prior}{A list containing the prior hyperparameters. See
\code{\link{nc_prior}()} for more details.}

\item{prm_init}{An object of class \code{list} containing parameters used
for initializing the simulation.}

\item{init}{Named \code{list} providing user-defined starting values.
If provided, it overrides the value given for \code{random_start}.}

\item{tuning}{Named \code{list} of tuning parameters (see details).}

\item{adaptation}{Named \code{list} of adaptation parameters (see details).}
}
\value{
\code{netcopula} returns an object of class \code{\link{nc_mcmc}}.
}
\description{
Estimation of a Copula-Based Model for Multivariate Network Meta-Analysis.
}
\examples{
\dontrun{
demo("example_homesafety", package = "netcopula")
}

}
\seealso{
\code{\link{nc_data-class}},
\code{\link{nc_mcmc}}
}
\author{
Sergio Venturini \email{sergio.venturini@unibocconi.it}
}
