% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_classes.R
\docType{methods}
\name{summary,nc_mcmc-method}
\alias{summary,nc_mcmc-method}
\title{Provide a summary of a \code{nc_mcmc} class instance.}
\usage{
\S4method{summary}{nc_mcmc}(object, include.burnin = FALSE,
  regex_pars = "d", ...)
}
\arguments{
\item{object}{An object of class \code{\link{nc_mcmc}}.}

\item{include.burnin}{A length-one logical vector. If \code{TRUE} the
burnin iterations (if available) are included in the summary.}

\item{regex_pars}{An optional \code{\link[=grep]{regular expression}} to use for
parameter selection.}

\item{...}{Further arguments to pass on (currently ignored).}
}
\description{
Provide a summary of a \code{nc_mcmc} class instance.
}
\author{
Sergio Venturini \email{sergio.venturini@unibocconi.it}
}
